Moved clean_ascwds_workplace_data into new folder structure. Moved associated tests, test data/schema. Changed glue script directory.
