+++ tests/test_file_data.py
@@ -219,254 +219,6 @@ class ASCWDSWorkerData:
 
 
 @dataclass
-class ASCWDSWorkplaceData:
-    workplace_rows = [
-        (
-            "1-000000001",
-            "101",
-            14,
-            16,
-            "20200101",
-            "1",
-            date(2021, 2, 1),
-            0,
-            "201",
-            "01/02/2021",
-            "A101",
-        ),
-        (
-            "1-000000002",
-            "102",
-            76,
-            65,
-            "20200101",
-            "1",
-            date(2021, 4, 1),
-            1,
-            None,
-            "01/02/2021",
-            "A102",
-        ),
-        (
-            "1-000000003",
-            "103",
-            34,
-            34,
-            "20200101",
-            "2",
-            date(2021, 3, 1),
-            0,
-            "203",
-            "01/02/2021",
-            "A103",
-        ),
-        (
-            "1-000000004",
-            "104",
-            234,
-            265,
-            "20190101",
-            "2",
-            date(2021, 4, 1),
-            0,
-            None,
-            "01/02/2021",
-            "A104",
-        ),
-        (
-            "1-000000005",
-            "105",
-            62,
-            65,
-            "20190101",
-            "3",
-            date(2021, 10, 1),
-            0,
-            None,
-            "01/02/2021",
-            "A105",
-        ),
-        (
-            "1-000000006",
-            "106",
-            77,
-            77,
-            "20190101",
-            "3",
-            date(2020, 3, 1),
-            1,
-            None,
-            "01/02/2021",
-            "A106",
-        ),
-        (
-            "1-000000007",
-            "107",
-            51,
-            42,
-            "20190101",
-            " 3",
-            date(2021, 5, 1),
-            0,
-            None,
-            "01/05/2021",
-            "A107",
-        ),
-        (
-            "1-000000008",
-            "108",
-            36,
-            34,
-            "20190101",
-            "4",
-            date(2021, 7, 1),
-            0,
-            None,
-            "01/05/2021",
-            "A108",
-        ),
-        (
-            "1-000000009",
-            "109",
-            34,
-            32,
-            "20190101",
-            "5",
-            date(2021, 12, 1),
-            0,
-            None,
-            "01/05/2021",
-            "A109",
-        ),
-        (
-            "1-0000000010",
-            "110",
-            14,
-            20,
-            "20190101",
-            "6",
-            date(2021, 3, 1),
-            0,
-            None,
-            "01/05/2021",
-            "A1010",
-        ),
-    ]
-
-    filter_test_account_when_orgid_present_rows = [
-        ("1-001", "310"),
-        ("1-002", "2452"),
-        ("1-003", "308"),
-        ("1-004", "1234"),
-        ("1-005", "31138"),
-    ]
-    expected_filter_test_account_when_orgid_present_rows = [
-        ("1-004", "1234"),
-    ]
-
-    filter_test_account_when_orgid_not_present_rows = [
-        ("1-001", "20250101"),
-        ("1-002", "20250101"),
-        ("1-003", "20250101"),
-        ("1-004", "20250101"),
-        ("1-005", "20250101"),
-    ]
-
-    remove_white_space_from_nmdsid_rows = [
-        ("1-001", "A123  "),
-        ("1-002", "A1234 "),
-        ("1-003", "A12345"),
-    ]
-    expected_remove_white_space_from_nmdsid_rows = [
-        ("1-001", "A123"),
-        ("1-002", "A1234"),
-        ("1-003", "A12345"),
-    ]
-
-    small_location_rows = [
-        ("loc-1", "2020-01-01", "1"),
-        ("loc-2", "2020-01-01", "2"),
-        ("loc-3", "2020-01-01", "3"),
-        ("loc-4", "2021-01-01", "4"),
-        (None, "2021-01-01", "5"),
-        (None, "2021-01-01", "6"),
-    ]
-
-    location_rows_with_duplicates = [
-        *small_location_rows,
-        ("loc-3", "2020-01-01", "7"),
-        ("loc-4", "2021-01-01", "8"),
-    ]
-
-    location_rows_with_different_import_dates = [
-        *small_location_rows,
-        ("loc-3", "2021-01-01", "3"),
-        ("loc-4", "2022-01-01", "4"),
-    ]
-
-    expected_filtered_location_rows = [
-        ("loc-1", "2020-01-01", "1"),
-        ("loc-2", "2020-01-01", "2"),
-        (None, "2021-01-01", "5"),
-        (None, "2021-01-01", "6"),
-    ]
-
-    mupddate_for_org_rows = [
-        ("1", date(2024, 3, 1), "1", date(2024, 1, 10)),
-        ("1", date(2024, 3, 1), "2", date(2024, 1, 20)),
-        ("1", date(2024, 4, 1), "3", date(2024, 3, 10)),
-        ("1", date(2024, 4, 1), "4", date(2024, 3, 15)),
-        ("2", date(2024, 4, 1), "5", date(2024, 2, 15)),
-        ("2", date(2024, 4, 1), "6", date(2024, 3, 10)),
-    ]
-    expected_mupddate_for_org_rows = [
-        ("1", date(2024, 3, 1), "1", date(2024, 1, 10), date(2024, 1, 20)),
-        ("1", date(2024, 3, 1), "2", date(2024, 1, 20), date(2024, 1, 20)),
-        ("1", date(2024, 4, 1), "3", date(2024, 3, 10), date(2024, 3, 15)),
-        ("1", date(2024, 4, 1), "4", date(2024, 3, 15), date(2024, 3, 15)),
-        ("2", date(2024, 4, 1), "5", date(2024, 2, 15), date(2024, 3, 10)),
-        ("2", date(2024, 4, 1), "6", date(2024, 3, 10), date(2024, 3, 10)),
-    ]
-
-    add_purge_data_col_rows = [
-        ("1", "Yes", date(2024, 2, 2), date(2024, 2, 2)),
-        ("2", "Yes", date(2024, 2, 2), date(2024, 3, 3)),
-        ("3", "No", date(2024, 2, 2), date(2024, 2, 2)),
-        ("4", "No", date(2024, 2, 2), date(2024, 3, 3)),
-    ]
-    expected_add_purge_data_col_rows = [
-        ("1", "Yes", date(2024, 2, 2), date(2024, 2, 2), date(2024, 2, 2)),
-        ("2", "Yes", date(2024, 2, 2), date(2024, 3, 3), date(2024, 3, 3)),
-        ("3", "No", date(2024, 2, 2), date(2024, 2, 2), date(2024, 2, 2)),
-        ("4", "No", date(2024, 2, 2), date(2024, 3, 3), date(2024, 2, 2)),
-    ]
-
-    add_workplace_last_active_date_col_rows = [
-        ("1", date(2024, 3, 3), date(2024, 2, 2)),
-        ("2", date(2024, 4, 4), date(2024, 5, 5)),
-    ]
-    expected_add_workplace_last_active_date_col_rows = [
-        ("1", date(2024, 3, 3), date(2024, 2, 2), date(2024, 3, 3)),
-        ("2", date(2024, 4, 4), date(2024, 5, 5), date(2024, 5, 5)),
-    ]
-
-    date_col_for_purging_rows = [
-        ("1", date(2024, 3, 3)),
-        ("2", date(2024, 4, 4)),
-    ]
-    expected_date_col_for_purging_rows = [
-        ("1", date(2024, 3, 3), date(2022, 3, 3)),
-        ("2", date(2024, 4, 4), date(2022, 4, 4)),
-    ]
-
-    workplace_last_active_rows = [
-        ("1", date(2024, 4, 4), date(2024, 5, 5)),
-        ("2", date(2024, 4, 4), date(2024, 4, 4)),
-        ("3", date(2024, 4, 4), date(2024, 3, 3)),
-    ]
-
-
-@dataclass
 class CQCProviderData:
     sample_rows_full = [
         (
